Convert OID based attributes back to friendly names
WAYF's code assumes that attributes will be passed to PHPH using thier
friendly name. This requires special provision in the hub, so it is
easier for us to send the default OID-based names. This change attempts
to detect OID-based attribute names and convert the back to their
friendly form, allowing either version to work. It was first used by
SAFIRE on 2016/08/25 and is backported from there.