Make oid2basic (more) deterministic
The original basic2oid had a one-to-one mapping between basic names and
oids. This allowed a simple array_flip to produce oid2basic. However, in
reality the "basic" names aren't so well defined, and there are several
versions of the same thing in use (e.g. givenName vs gn). This makes it
desirable to have a many-to-one mapping in basic2oid but still requires
a one-to-one mapping in oid2basic (there must be a canonical form).